src/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-autoslug==1.9.9
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3